pyproject.toml
[project]
name = "client-connector"
version = "0.1.0"
description = "MCP Host Connector - Gateway for AI agents to connect to ConHub infrastructure"
readme = "README.md"
requires-python = ">=3.14"
license = {text = "MIT"}

dependencies = [
    # Web framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    
    # MCP Server
    "fastmcp>=0.2.0",
    
    # WebSocket support
    "websockets>=12.0",
    
    # HTTP client
    "httpx>=0.26.0",
    
    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    
    # Data validation
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Database
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.25",
    
    # Async utilities
    "anyio>=4.2.0",
    
    # Logging
    "structlog>=24.1.0",
    
    # Environment
    "python-dotenv>=1.0.0",
    "uuid6>=2024.1.0",
    "redis>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
    "fakeredis>=2.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ["py314"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
